Historic Hash, Intermediate Save Dublicates

+++ collectmeterdigits/__main__.py
@@ -11,7 +11,7 @@ def main():
     parser.add_argument('--labeling', default='', help='labelpath if you want label the images')
     parser.add_argument('--keepdownloads', action='store_true', help='Normally all downloaded data will be deleted. If set it keeps the images.')
     parser.add_argument('--nodownload', action='store_true', help='Do not download pictures. Only remove duplicates and labeling.')
-    parser.add_argument('--startlabel', type=int, default=0, help='only images >= startlabel. (default: all)')
+    parser.add_argument('--startlabel', type=float, default=0.0, help='only images >= startlabel. (default: all)')
 
     # print help message if no argument is given
     if len(sys.argv)==1:
